+++ test/printly_test.dart
@@ -17,7 +17,10 @@ class _MockPrintlyPlatform extends PrintlyPlatform
   Future<bool> openBluetoothSettings() async => true;
 
   @override
-  Future<void> startScan({required Set<ConnectionType> types}) async {}
+  Future<void> startScan({
+    required Set<ConnectionType> types,
+    bool includeUnnamed = false,
+  }) async {}
 
   @override
   Future<void> stopScan() async {}